extract_function/requirements.txt
requests==2.28.2
beautifulsoup4==4.11.2
functions-framework==3.3.0
google-cloud-storage==2.7.0
